2012/6/9 PLW_DemoV4(two visual demos).zip
Hi,
If there is no some flaw that I have mistakenly over-passed, I think the
visual stimuli is (almost %just in case there is something else about it)
ready to use. And you can run it.
ClockPLW.m & RL_PLW.m are main scripts where they stand for PLWs on the
surface of a circle like a clock & PLWs overlapping right and left.

As for the experiments:
ClockPLW.m :
every trial the PLWs are regenerated randomly in the direction they are
facing.

RL_PLW.m   :
1. Two PLWs either both upright or both inverted vertically while the facing
direction of each of them is random. ( So the red one is not necessarily
should face in the opposite direction with the green one. )
2. Each PLWs walking pace is random in three levels.
3. Starting position of each PLW is also random.

At last, there are no more the so called "looping issue : ( " now!  : D
)